storybook: 기본 컴포넌트 제거 및 Home 스토리북 추가
